+++ CHANGELOG.md
@@ -6,6 +6,7 @@
 ### Fixed
 
 - Fixed a bug where ignition wouldn't work from the PHAR archive
+- Fixed a missing dependency registered as a dev-dependency
 
 2.0.0 - 2014-09-17
 ------------------