+++ CHANGELOG.md
@@ -11,7 +11,8 @@
 ### Fixed
 - Fixed a bug where ignition wouldn't work from the PHAR archive
 - Fixed a missing dependency registered as a dev-dependency
-- Fixed strictness of which system that would fail on some binaries
+- Fixed strictness of `which` system that would fail on some binaries
+- Fixed a bug where events would never be registered when using only `hooks.php` with one connection/stage
 
 2.0.0 - 2014-09-17
 ------------------